requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0

SQLAlchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
python-dotenv==1.1.1

argon2-cffi==23.1.0
python-jose[cryptography]==3.5.0
pydantic==2.11.7
pydantic-settings==2.10.1

pydantic[email]

python-multipart==0.0.9
boto3==1.35.70